[#7] rename: 영수증 페이지 폴더명 변경
