+++ src/pages/ReceiptPage/ReceiptPage.jsx
