+++ src/router/Router.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import MyPage from '../pages/MyPage/MyPage';
 import LoginPage from '../pages/LoginPage/LoginPage';
 import RecordPage from '../pages/RecordPage/RedcordPage';
-import UserSharePage from '../pages/UserSharePage/UserSharePage';
+import ReceiptPage from '../pages/ReceiptPage/ReceiptPage';
 
 export default function Router() {
   return (
@@ -12,7 +12,7 @@ export default function Router() {
         <Route path='/' element={<MyPage />} />
         <Route path='/login' element={<LoginPage />} />
         <Route path='/recordPage' element={<RecordPage />} />
-        <Route path='/share' element={<UserSharePage />} />
+        <Route path='/receipt' element={<ReceiptPage />} />
       </Routes>
     </BrowserRouter>
   );
